requirements.txt
fastapi
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
email-validator
uvicorn